+++ src/main/scala/traffic/countsTree.scala
@@ -1,7 +1,7 @@
 package traffic
 
 
-case class CountsTree(label: String, count: Int, children: Iterable[CountsTree]) {
+case class CountsTree(label: String, count: Long, children: Iterable[CountsTree]) {
   def print(indent: Int): Unit = {
     println(s">${"  " * indent}${label}: ${count}")
     children.foreach{ _.print(indent + 1) }
